+++ internal/provider/dashboards/dashboard_schema/v4.go
@@ -1289,6 +1289,7 @@ func dashboardSchemaAttributesV4() map[string]schema.Attribute {
 									objectvalidator.ExactlyOneOf(
 										path.MatchRelative().AtParent().AtName("logs"),
 										path.MatchRelative().AtParent().AtName("spans"),
+										path.MatchRelative().AtParent().AtName("manual"),
 									),
 								},
 							},
@@ -1299,6 +1300,7 @@ func dashboardSchemaAttributesV4() map[string]schema.Attribute {
 									objectvalidator.ExactlyOneOf(
 										path.MatchRelative().AtParent().AtName("metrics"),
 										path.MatchRelative().AtParent().AtName("spans"),
+										path.MatchRelative().AtParent().AtName("manual"),
 									),
 								},
 							},
@@ -1309,9 +1311,11 @@ func dashboardSchemaAttributesV4() map[string]schema.Attribute {
 									objectvalidator.ExactlyOneOf(
 										path.MatchRelative().AtParent().AtName("metrics"),
 										path.MatchRelative().AtParent().AtName("logs"),
+										path.MatchRelative().AtParent().AtName("manual"),
 									),
 								},
 							},
+							"manual": manualAnnotationSourceAttribute(),
 						},
 						Required: true,
 					},